Change file naming style from underscore to dash
