+++ test/data_channel_message_queue.test.js
@@ -1,5 +1,5 @@
-import DataChannelMessage from '../src/data_channel_message';
-import DataChannelMessageQueue from '../src/data_channel_message_queue';
+import DataChannelMessage from '../src/data-channel-message';
+import DataChannelMessageQueue from '../src/data-channel-message-queue';
 import { WEBRTC_DATACHANNEL_CHUNK_SIZE } from '../src';
 import { randomFillSync } from 'crypto';
